embedding/requirements.txt
faiss-cpu==1.7.4
numpy==2.4.0
openai==2.14.0
python-dotenv==1.1.1
